api/requirements.txt
fastapi[standard]>=0.115.11
google-cloud-firestore>=2.20.1
google-cloud-run>=0.10.16
pydantic-settings>=2.8.1
uvicorn>=0.34.0